+++ R/translate-sql-base.r
@@ -66,7 +66,7 @@ base_scalar <- sql_translator(
   exp     = sql_prefix("exp", 1),
   floor   = sql_prefix("floor", 1),
   log     = function(x, base = exp(1)) {
-    build_sql(sql("log"), list(x, base))
+    build_sql(sql("log"), list(base, x))
   },
   log10   = sql_prefix("log10", 1),
   round   = sql_prefix("round", 2),